requirements.txt
apache-airflow==2.9.1
json5==0.9.10
pandas==2.0.3